budgetFactor should be a float not a double
git-svn-id: file:///semplestb/final_repo@3265 02200ff9-b5b2-46f0-9171-221b09c08c7b

+++ SemplestServices/src/semplest/service/bidding/BidGeneratorObj.java
@@ -75,7 +75,7 @@ public class BidGeneratorObj
 	private Long stepRest;// = 800000L;
 	
 	Double googlePercent;
-	Double budgetFactor;
+	Float budgetFactor;
 	
 
 	private String googleAccountID = null;
@@ -159,7 +159,7 @@ public class BidGeneratorObj
 		
 		final Integer googlePercentInteger = (Integer)SemplestConfiguration.configData.get("SemplestBiddingGooglePercent");		
 		googlePercent =  Double.valueOf("" + googlePercentInteger);
-		budgetFactor = (Double)SemplestConfiguration.configData.get("SemplestBiddingBudgetMultFactor");
+		budgetFactor = (Float)SemplestConfiguration.configData.get("SemplestBiddingBudgetMultFactor");
 		
 		
 		final Double budgetBoostFactorDouble = (Double) SemplestConfiguration.configData.get("SemplestBiddingInitialBidBoostFactor");
